+++ TemplateGenerator.sln
@@ -9,7 +9,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{ABBEC3FA
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TemplateCreator.Tests", "tests\TemplateCreator.Tests\TemplateCreator.Tests.csproj", "{D24A5FB9-BE24-4181-BA7C-980A63DBE6AC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TemplateCreator", "src\TemplateCreator\TemplateCreator\TemplateCreator\TemplateCreator.csproj", "{E20AD4D1-883D-492A-AE8B-8B8A74776E36}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TemplateCreator", "src\TemplateCreator\TemplateCreator.csproj", "{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -45,33 +45,33 @@ Global
 		{D24A5FB9-BE24-4181-BA7C-980A63DBE6AC}.Release|x64.Build.0 = Release|Any CPU
 		{D24A5FB9-BE24-4181-BA7C-980A63DBE6AC}.Release|x86.ActiveCfg = Release|Any CPU
 		{D24A5FB9-BE24-4181-BA7C-980A63DBE6AC}.Release|x86.Build.0 = Release|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Debug|ARM.Build.0 = Debug|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Debug|ARM64.ActiveCfg = Debug|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Debug|ARM64.Build.0 = Debug|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Debug|x64.Build.0 = Debug|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Debug|x86.Build.0 = Debug|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Release|ARM.ActiveCfg = Release|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Release|ARM.Build.0 = Release|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Release|ARM64.ActiveCfg = Release|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Release|ARM64.Build.0 = Release|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Release|x64.ActiveCfg = Release|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Release|x64.Build.0 = Release|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Release|x86.ActiveCfg = Release|Any CPU
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36}.Release|x86.Build.0 = Release|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Debug|ARM.Build.0 = Debug|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Debug|ARM64.ActiveCfg = Debug|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Debug|ARM64.Build.0 = Debug|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Debug|x64.Build.0 = Debug|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Debug|x86.Build.0 = Debug|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Release|ARM.ActiveCfg = Release|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Release|ARM.Build.0 = Release|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Release|ARM64.ActiveCfg = Release|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Release|ARM64.Build.0 = Release|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Release|x64.ActiveCfg = Release|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Release|x64.Build.0 = Release|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Release|x86.ActiveCfg = Release|Any CPU
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{D24A5FB9-BE24-4181-BA7C-980A63DBE6AC} = {ABBEC3FA-3791-4522-B0AC-911A73AD68F6}
-		{E20AD4D1-883D-492A-AE8B-8B8A74776E36} = {2D2F0F40-EDE1-4DB7-947C-5694354A1A90}
+		{5F471582-EEEC-4F8D-8DFA-DDCDC7D2520D} = {2D2F0F40-EDE1-4DB7-947C-5694354A1A90}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {297B7372-B463-4F3B-AC9A-7C4BA19BDD91}
