+++ src/TemplateCreator/MainWindow.xaml
@@ -1,0 +1,25 @@
+﻿<Window 
+        Style="{StaticResource MaterialDesignWindow}"
+	    x:Class="TemplateCreator.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:TemplateCreator"
+        xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes" 
+	    FontFamily="{materialDesign:MaterialDesignFont}"
+        TextElement.Foreground="{DynamicResource MaterialDesign.Brush.Foreground}"
+        Background="{DynamicResource MaterialDesign.Brush.Background}"
+        mc:Ignorable="d"
+        Title="MainWindow" Height="450" Width="800">
+	<StackPanel Margin="5,5,5,5">
+		<StackPanel Orientation="Horizontal">
+			<ComboBox
+				Text="Select Process"
+				ItemsSource="{Binding Processes}"
+                SelectedItem="{Binding SelectedProcess}" 
+                DisplayMemberPath="ProcessName"/>
+			<Button Margin="5,5,5,5" Content="Screenshot" IsEnabled="{Binding IsProcessSelected}" />
+		</StackPanel>
+	</StackPanel>
+</Window>
