--- src/TemplateCreator/TemplateCreator/Directory.Build.props
@@ -1,32 +1,0 @@
-<!-- 
-This file allow for customizing your build process.
-See: https://learn.microsoft.com/visualstudio/msbuild/customize-your-build
--->
-<Project>
-  <!--
-  Uncomment if you need to enable inclusion of another Directory.Build.props file from a parent directory
-  <Import Project="$([MSBuild]::GetPathOfFileAbove('Directory.Build.props', '$(MSBuildThisFileDirectory)../'))" />
-  -->
-
-  <PropertyGroup>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <LangVersion>12</LangVersion>
-    <AccelerateBuildsInVisualStudio>true</AccelerateBuildsInVisualStudio>
-    <!-- 
-    If you you like to see source generated files saved to disk you can enable the following:
-    https://learn.microsoft.com/dotnet/csharp/roslyn-sdk/source-generators-overview?WT.mc_id=DT-MVP-5003472
-    -->
-    <!--<EmitCompilerGeneratedFiles>true</EmitCompilerGeneratedFiles>-->
-  </PropertyGroup>
-
-  <!--
-  This allows all projects to share the same user secrets file. 
-  If you want project to have their own, set it to a different GUID on each project.
-  See: https://learn.microsoft.com/dotnet/architecture/microservices/secure-net-microservices-web-applications/developer-app-secrets-storage
-  -->
-  <PropertyGroup Label="User Secrets">
-    <UserSecretsId>6df35fb8-f595-4c52-bfb7-5e7cd2e55c45</UserSecretsId>
-  </PropertyGroup>
-
-</Project>
