--- src/TemplateCreator/TemplateCreator/NuGet.config
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <!-- `clear` ensures no additional sources are inherited from another config file. -->
-  <packageSources>
-    <clear />
-    <!-- `key` can be any identifier for your source. -->
-    <add key="nuget.org" value="https://api.nuget.org/v3/index.json" />
-  </packageSources>
-
-  <!-- Define mappings by adding package patterns beneath the target source. -->
-  <packageSourceMapping>
-    <!-- key value for <packageSource> should match key values from <packageSources> element -->
-    <packageSource key="nuget.org">
-      <package pattern="*" />
-    </packageSource>
-  </packageSourceMapping>
-</configuration>
-
-<!--
-As an example, you can define multiple package sources and map particular nuget packages from each source.
-This defines the package sources, nuget.org and contoso.com and maps all NuGet packages with Contoso.* to it.
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <packageSources>
-    <clear />
-    <add key="nuget.org" value="https://api.nuget.org/v3/index.json" />
-    <add key="contoso.com" value="https://contoso.com/packages/" />
-  </packageSources>
-  
-  <packageSourceMapping>
-    <packageSource key="nuget.org">
-      <package pattern="*" />
-    </packageSource>
-    <packageSource key="contoso.com">
-      <package pattern="Contoso.*" />
-    </packageSource>
-  </packageSourceMapping>
-</configuration>
--->
