--- src/TemplateCreator/TemplateCreator/README.md
@@ -1,35 +1,0 @@
-# WPF app template
-This template creates a full WPF application, along with unit tests.
-
-## Template
-Create a new app in your current directory by running.
-
-```cli
-> dotnet new keboo.wpf
-```
-
-### Parameters
-[Default template options](https://learn.microsoft.com/dotnet/core/tools/dotnet-new#options)
-
-## Key Features
-
-### Generic Host Dependency Injection
-*TODO*
-
-### CEntralized Package Management
-*TODO*
-
-### Build Customization
-*TODO*
-
-### CommunityToolkit MVVM
-*TODO*
-
-### Material Design in XAML
-*TODO*
-
-### .editorconfig formatting
-*TODO*
-
-### Testing with Moq.AutoMocker
-*TODO*
