--- src/TemplateCreator/TemplateCreator/Settings.XamlStyler
@@ -1,44 +1,0 @@
-{
-    "IndentSize": 2,
-    "AttributesTolerance": 2,
-    "KeepFirstAttributeOnSameLine": true,
-    "MaxAttributeCharactersPerLine": 0,
-    "MaxAttributesPerLine": 1,
-    "NewlineExemptionElements": "RadialGradientBrush, GradientStop, LinearGradientBrush, ScaleTransform, SkewTransform, RotateTransform, TranslateTransform, Trigger, MultiTrigger, DataTrigger, MultiDataTrigger, Condition, Setter, SolidColorBrush, Pen",
-    "SeparateByGroups": false,
-    "AttributeIndentation": 0,
-    "AttributeIndentationStyle": "Spaces",
-    "RemoveDesignTimeReferences": false,
-    "EnableAttributeReordering": true,
-    "AttributeOrderingRuleGroups": [
-        "x:Class",
-        "xmlns, xmlns:x",
-        "xmlns:*",
-        "x:Key, Key, x:Name, Name, x:Uid, Uid, Title",
-        "Grid.Row, Grid.RowSpan, Grid.Column, Grid.ColumnSpan, Canvas.Left, Canvas.Top, Canvas.Right, Canvas.Bottom",
-        "Width, Height, MinWidth, MinHeight, MaxWidth, MaxHeight",
-        "Margin, Padding, HorizontalAlignment, VerticalAlignment, HorizontalContentAlignment, VerticalContentAlignment, Panel.ZIndex",
-        "*:*, *",
-        "PageSource, PageIndex, Offset, Color, TargetName, Property, Value, StartPoint, EndPoint",
-        "*:Freeze, mc:Ignorable, d:IsDataSource, d:LayoutOverrides, d:IsStaticText",
-        "Storyboard.*, From, To, Duration",
-        "TargetType",
-        "BasedOn"
-    ],
-    "FirstLineAttributes": "",
-    "OrderAttributesByName": true,
-    "PutEndingBracketOnNewLine": false,
-    "RemoveEndingTagOfEmptyElement": true,
-    "SpaceBeforeClosingSlash": true,
-    "RootElementLineBreakRule": "Always",
-    "ReorderVSM": "First",
-    "ReorderGridChildren": false,
-    "ReorderCanvasChildren": false,
-    "ReorderSetters": "TargetNameThenProperty",
-    "FormatMarkupExtension": true,
-    "NoNewLineMarkupExtensions": "x:Bind, Binding, TemplateBinding, x:Static, DynamicResource, ComponentResourceKey, iconPacks:Modern, iconPacks:Material",
-    "ThicknessSeparator": "Comma",
-    "ThicknessAttributes": "Margin, Padding, BorderThickness, ThumbnailClipMargin",
-    "FormatOnSave": false,
-    "CommentPadding": 1
-}
