--- src/TemplateCreator/TemplateCreator/TemplateCreator/App.xaml
@@ -1,19 +1,0 @@
-﻿<Application x:Class="TemplateCreator.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:TemplateCreator"
-             xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes">
-  <Application.Resources>
-    <ResourceDictionary>
-      <ResourceDictionary.MergedDictionaries>
-        <materialDesign:BundledTheme BaseTheme="Inherit" 
-                                     PrimaryColor="Purple" 
-                                     SecondaryColor="LightGreen" />
-
-        <ResourceDictionary
-                    Source="pack://application:,,,/MaterialDesignThemes.Wpf;component/Themes/MaterialDesign3.Defaults.xaml" />
-
-      </ResourceDictionary.MergedDictionaries>
-    </ResourceDictionary>
-  </Application.Resources>
-</Application>
