--- src/TemplateCreator/TemplateCreator/TemplateCreator/MainWindow.xaml
@@ -1,32 +1,0 @@
-﻿<Window x:Class="TemplateCreator.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:TemplateCreator"
-        mc:Ignorable="d"
-        d:DataContext="{d:DesignInstance Type=local:MainWindowViewModel, IsDesignTimeCreatable=False}"
-        Style="{StaticResource MaterialDesignWindow}"
-        Title="Demo App" Height="450" Width="800">
-  <Grid>
-    <Grid.RowDefinitions>
-      <RowDefinition Height="Auto"/>
-      <RowDefinition />
-      <RowDefinition Height="Auto"/>
-    </Grid.RowDefinitions>
-
-    <Menu>
-      <MenuItem Header="_File">
-        <MenuItem Header="_Clear Count" Command="{Binding ClearCountCommand}"/>
-        <Separator />
-        <MenuItem Header="E_xit" Command="{x:Static ApplicationCommands.Close}"/>
-      </MenuItem>
-    </Menu>
-    
-    <StackPanel Grid.Row="1" HorizontalAlignment="Center" VerticalAlignment="Center">
-      <Image />
-      <TextBlock Text="{Binding Count, StringFormat='{}{0} Clicks'}" Margin="0,10"/>
-      <Button Content="_Click Me" Command="{Binding IncrementCountCommand}" />
-    </StackPanel>
-  </Grid>
-</Window>
