--- src/TemplateCreator/TemplateCreator/TemplateCreator/MainWindow.xaml.cs
@@ -1,23 +1,0 @@
-﻿using System.Windows.Input;
-
-namespace TemplateCreator
-{
-    /// <summary>
-    /// Interaction logic for MainWindow.xaml
-    /// </summary>
-    public partial class MainWindow
-    {
-        public MainWindow(MainWindowViewModel viewModel)
-        {
-            DataContext = viewModel;
-            InitializeComponent();
-
-            CommandBindings.Add(new CommandBinding(ApplicationCommands.Close, OnClose));
-        }
-
-        private void OnClose(object sender, ExecutedRoutedEventArgs e)
-        {
-            Close();
-        }
-    }
-}
